requirement.txt
fastapi
python-dotenv
requests
aiohttp
